chore: Drop sha/hmac custom deps to RustCrypto + drop boringssl
