+++ Cargo.toml
@@ -11,16 +11,21 @@ categories = ["wasm", "api-bindings", "web-programming", "cryptography"]
 keywords = ["jwt", "jws", "authentication", "tokens", "json"]
 readme = "README.md"
 
+[features]
+cwt = ["ciborium"]
+rsa = ["dep:rsa"]
+
 [dependencies]
 anyhow = "1.0.75"
 binstring = "0.1.1"
 ciborium = { version = "0.2.1", optional = true }
 coarsetime = "0.1.33"
 ct-codecs = "1.1.1"
-ed25519-compact = { version = "2.0.4", features = ["pem"] }
-hmac-sha1-compact = "1.1.4"
-hmac-sha256 = { version = "1.1.7", features = ["traits010"] }
-hmac-sha512 = { version = "1.1.5", features = ["traits010", "sha384"] }
+ed25519-compact = { version = "2.1", features = ["pem"] }
+hmac = "0.12"
+digest = "0.10"
+sha-1 = { version = "0.10", features = ["oid"] }
+sha2 = { version = "0.10", features = ["oid"] }
 ecdsa = { version = "0.16", features = ["der"] }
 k256 = { version = "0.13.2", features = ["ecdsa", "std", "pkcs8", "pem"] }
 p256 = { version = "0.13.2", features = ["ecdsa", "std", "pkcs8", "pem"] }
@@ -32,20 +37,12 @@ serde_json = "1.0.108"
 thiserror = "1.0.50"
 zeroize = "1.7.0"
 getrandom = { version = "0.2", features = ["js"] }
+rsa = { version = "0.9", optional = true }
+spki = "0.7"
 
-[target.'cfg(not(any(target_arch = "wasm32", target_arch = "wasm64")))'.dependencies]
-boring = { version = "4.1.0", optional = true }
-
-[target.'cfg(any(target_arch = "wasm32", target_arch = "wasm64"))'.dependencies]
-rsa = { version = "0.8.2", optional = true }
-spki = "0.6.0"
 
 [dev-dependencies]
 benchmark-simple = "0.1.8"
-
-[features]
-cwt = ["ciborium"]
-rsa = ["dep:rsa", "dep:boring"]
 
 [[bench]]
 name = "benchmark"
