+++ src/algorithms/es256k.rs
@@ -1,3 +1,4 @@
+use digest::Digest;
 use std::convert::TryFrom;
 
 use ct_codecs::{Base64UrlSafeNoPadding, Encoder};
@@ -165,7 +166,7 @@ pub trait ECDSAP256kKeyPairLike {
     ) -> Result<String, Error> {
         let jwt_header = header.with_metadata(self.metadata());
         Token::build(&jwt_header, Some(claims), |authenticated| {
-            let mut digest = hmac_sha256::Hash::new();
+            let mut digest = sha2::Sha256::new();
             digest.update(authenticated.as_bytes());
             let mut rng = rand::thread_rng();
             let signature: ecdsa::Signature = self
@@ -195,7 +196,7 @@ pub trait ECDSAP256kPublicKeyLike {
             |authenticated, signature| {
                 let ecdsa_signature = ecdsa::Signature::try_from(signature)
                     .map_err(|_| JWTError::InvalidSignature)?;
-                let mut digest = hmac_sha256::Hash::new();
+                let mut digest = sha2::Sha256::new();
                 digest.update(authenticated.as_bytes());
                 self.public_key()
                     .as_ref()
@@ -219,7 +220,7 @@ pub trait ECDSAP256kPublicKeyLike {
             |authenticated, signature| {
                 let ecdsa_signature = ecdsa::Signature::try_from(signature)
                     .map_err(|_| JWTError::InvalidSignature)?;
-                let mut digest = hmac_sha256::Hash::new();
+                let mut digest = sha2::Sha256::new();
                 digest.update(authenticated.as_bytes());
                 self.public_key()
                     .as_ref()
@@ -232,7 +233,7 @@ pub trait ECDSAP256kPublicKeyLike {
 
     fn create_key_id(&mut self) -> &str {
         self.set_key_id(
-            Base64UrlSafeNoPadding::encode_to_string(hmac_sha256::Hash::hash(
+            Base64UrlSafeNoPadding::encode_to_string(sha2::Sha256::digest(
                 &self.public_key().to_bytes(),
             ))
             .unwrap(),
