+++ src/algorithms/mod.rs
@@ -5,13 +5,8 @@ mod es384;
 mod es512;
 mod hmac;
 
-#[cfg(all(
-    feature = "rsa",
-    not(any(target_arch = "wasm32", target_arch = "wasm64"))
-))]
+#[cfg(feature = "rsa")]
 mod rsa;
-#[cfg(all(feature = "rsa", any(target_arch = "wasm32", target_arch = "wasm64")))]
-mod rsa_legacy;
 
 pub use self::eddsa::*;
 pub use self::es256::*;
@@ -20,10 +15,5 @@ pub use self::es384::*;
 pub use self::es512::*;
 pub use self::hmac::*;
 
-#[cfg(all(
-    feature = "rsa",
-    not(any(target_arch = "wasm32", target_arch = "wasm64"))
-))]
+#[cfg(feature = "rsa")]
 pub use self::rsa::*;
-#[cfg(all(feature = "rsa", any(target_arch = "wasm32", target_arch = "wasm64")))]
-pub use self::rsa_legacy::*;
