+++ src/algorithms/rsa.rs
@@ -1,12 +1,17 @@
-use boring::bn::BigNum;
-use boring::hash::MessageDigest;
-use boring::pkey::{PKey, Private, Public};
-use boring::rsa::{Padding, Rsa};
-use boring::sign::{Signer, Verifier};
+use std::mem;
+
 use ct_codecs::{Base64UrlSafeNoPadding, Encoder};
-use hmac_sha1_compact::Hash as SHA1;
-use hmac_sha256::Hash as SHA256;
+use digest::Digest;
+
+use rsa::{
+    pkcs1::{DecodeRsaPrivateKey as _, DecodeRsaPublicKey},
+    pkcs8::{DecodePrivateKey as _, DecodePublicKey as _, EncodePrivateKey as _},
+    traits::{PublicKeyParts as _, SignatureScheme},
+    BigUint,
+};
 use serde::{de::DeserializeOwned, Serialize};
+#[allow(unused_imports)]
+use spki::{DecodePublicKey as _, EncodePublicKey as _};
 
 use crate::claims::*;
 use crate::common::*;
@@ -18,10 +23,10 @@ use crate::token::*;
 
 #[doc(hidden)]
 #[derive(Debug, Clone)]
-pub struct RSAPublicKey(Rsa<Public>);
+pub struct RSAPublicKey(rsa::RsaPublicKey);
 
-impl AsRef<Rsa<Public>> for RSAPublicKey {
-    fn as_ref(&self) -> &Rsa<Public> {
+impl AsRef<rsa::RsaPublicKey> for RSAPublicKey {
+    fn as_ref(&self) -> &rsa::RsaPublicKey {
         &self.0
     }
 }
@@ -33,38 +38,41 @@ pub struct RSAPublicKeyComponents {
 
 impl RSAPublicKey {
     pub fn from_der(der: &[u8]) -> Result<Self, Error> {
-        let rsa_pk = Rsa::<Public>::public_key_from_der(der)
-            .or_else(|_| Rsa::<Public>::public_key_from_der_pkcs1(der))?;
+        let rsa_pk = rsa::RsaPublicKey::from_public_key_der(der)
+            .or_else(|_| rsa::RsaPublicKey::from_pkcs1_der(der))?;
         Ok(RSAPublicKey(rsa_pk))
     }
 
     pub fn from_pem(pem: &str) -> Result<Self, Error> {
         let pem = pem.trim();
-        let rsa_pk = Rsa::<Public>::public_key_from_pem(pem.as_bytes())
-            .or_else(|_| Rsa::<Public>::public_key_from_pem_pkcs1(pem.as_bytes()))?;
+        let rsa_pk = rsa::RsaPublicKey::from_public_key_pem(pem)
+            .or_else(|_| rsa::RsaPublicKey::from_pkcs1_pem(pem))?;
         Ok(RSAPublicKey(rsa_pk))
     }
 
     pub fn from_components(n: &[u8], e: &[u8]) -> Result<Self, Error> {
-        let n = BigNum::from_slice(n)?;
-        let e = BigNum::from_slice(e)?;
-        let rsa_pk = Rsa::<Public>::from_public_components(n, e)?;
+        let n = BigUint::from_bytes_be(n);
+        let e = BigUint::from_bytes_be(e);
+        let rsa_pk = rsa::RsaPublicKey::new(n, e)?;
         Ok(RSAPublicKey(rsa_pk))
     }
 
     pub fn to_der(&self) -> Result<Vec<u8>, Error> {
-        self.0.public_key_to_der().map_err(Into::into)
+        self.0
+            .to_public_key_der()
+            .map_err(Into::into)
+            .map(|x| x.as_ref().to_vec())
     }
 
     pub fn to_pem(&self) -> Result<String, Error> {
-        let bytes = self.0.public_key_to_pem()?;
-        let pem = String::from_utf8(bytes)?;
-        Ok(pem)
+        self.0
+            .to_public_key_pem(Default::default())
+            .map_err(Into::into)
     }
 
     pub fn to_components(&self) -> RSAPublicKeyComponents {
-        let n = self.0.n().to_vec();
-        let e = self.0.e().to_vec();
+        let n = self.0.n().to_bytes_be();
+        let e = self.0.e().to_bytes_be();
         RSAPublicKeyComponents { n, e }
     }
 }
@@ -72,22 +80,22 @@ impl RSAPublicKey {
 #[doc(hidden)]
 #[derive(Debug, Clone)]
 pub struct RSAKeyPair {
-    rsa_sk: Rsa<Private>,
+    rsa_sk: rsa::RsaPrivateKey,
     metadata: Option<KeyMetadata>,
 }
 
-impl AsRef<Rsa<Private>> for RSAKeyPair {
-    fn as_ref(&self) -> &Rsa<Private> {
+impl AsRef<rsa::RsaPrivateKey> for RSAKeyPair {
+    fn as_ref(&self) -> &rsa::RsaPrivateKey {
         &self.rsa_sk
     }
 }
 
 impl RSAKeyPair {
     pub fn from_der(der: &[u8]) -> Result<Self, Error> {
-        let rsa_sk = Rsa::<Private>::private_key_from_der(der)?;
-        if !(rsa_sk.check_key()?) {
-            bail!(JWTError::InvalidKeyPair);
-        }
+        let mut rsa_sk = rsa::RsaPrivateKey::from_pkcs8_der(der)
+            .or_else(|_| rsa::RsaPrivateKey::from_pkcs1_der(der))?;
+        rsa_sk.validate()?;
+        rsa_sk.precompute()?;
         Ok(RSAKeyPair {
             rsa_sk,
             metadata: None,
@@ -96,10 +104,10 @@ impl RSAKeyPair {
 
     pub fn from_pem(pem: &str) -> Result<Self, Error> {
         let pem = pem.trim();
-        let rsa_sk = Rsa::<Private>::private_key_from_pem(pem.as_bytes())?;
-        if !(rsa_sk.check_key()?) {
-            bail!(JWTError::InvalidKeyPair);
-        }
+        let mut rsa_sk = rsa::RsaPrivateKey::from_pkcs8_pem(pem)
+            .or_else(|_| rsa::RsaPrivateKey::from_pkcs1_pem(pem))?;
+        rsa_sk.validate()?;
+        rsa_sk.precompute()?;
         Ok(RSAKeyPair {
             rsa_sk,
             metadata: None,
@@ -107,27 +115,21 @@ impl RSAKeyPair {
     }
 
     pub fn to_der(&self) -> Result<Vec<u8>, Error> {
-        self.rsa_sk.private_key_to_der().map_err(Into::into)
+        self.rsa_sk
+            .to_pkcs8_der()
+            .map_err(Into::into)
+            .map(|x| mem::take(x.to_bytes().as_mut()))
     }
 
     pub fn to_pem(&self) -> Result<String, Error> {
-        let bytes = self.rsa_sk.private_key_to_pem()?;
-        let pem = String::from_utf8(bytes)?;
-        Ok(pem)
+        self.rsa_sk
+            .to_pkcs8_pem(Default::default())
+            .map_err(Into::into)
+            .map(|x| x.to_string())
     }
 
     pub fn public_key(&self) -> RSAPublicKey {
-        let rsa_pk = Rsa::<Public>::from_public_components(
-            self.rsa_sk
-                .n()
-                .to_owned()
-                .expect("failed to create public key"),
-            self.rsa_sk
-                .e()
-                .to_owned()
-                .expect("failed to create public key"),
-        )
-        .expect("failed to create public key");
+        let rsa_pk = self.rsa_sk.to_public_key();
         RSAPublicKey(rsa_pk)
     }
 
@@ -136,7 +138,8 @@ impl RSAKeyPair {
             2048 | 3072 | 4096 => {}
             _ => bail!(JWTError::UnsupportedRSAModulus),
         };
-        let rsa_sk = Rsa::<Private>::generate(modulus_bits as _)?;
+        let mut rng = rand::thread_rng();
+        let rsa_sk = rsa::RsaPrivateKey::new(&mut rng, modulus_bits)?;
         Ok(RSAKeyPair {
             rsa_sk,
             metadata: None,
@@ -145,13 +148,15 @@ impl RSAKeyPair {
 }
 
 pub trait RSAKeyPairLike {
+    type PaddingScheme: SignatureScheme;
+
     fn jwt_alg_name() -> &'static str;
     fn key_pair(&self) -> &RSAKeyPair;
     fn key_id(&self) -> &Option<String>;
     fn metadata(&self) -> &Option<KeyMetadata>;
     fn attach_metadata(&mut self, metadata: KeyMetadata) -> Result<(), Error>;
-    fn hash() -> MessageDigest;
-    fn padding_scheme(&self) -> Padding;
+    fn hash(message: &[u8]) -> Vec<u8>;
+    fn padding_scheme(&self) -> Self::PaddingScheme;
 
     fn sign<CustomClaims: Serialize + DeserializeOwned>(
         &self,
@@ -168,23 +173,27 @@ pub trait RSAKeyPairLike {
     ) -> Result<String, Error> {
         let jwt_header = header.with_metadata(self.metadata());
         Token::build(&jwt_header, Some(claims), |authenticated| {
-            let digest = Self::hash();
-            let pkey = PKey::from_rsa(self.key_pair().as_ref().clone())?;
-            let mut signer = Signer::new(digest, &pkey).unwrap();
-            signer.update(authenticated.as_bytes())?;
-            let signature = signer.sign_to_vec()?;
-            Ok(signature)
+            let digest = Self::hash(authenticated.as_bytes());
+            let mut rng = rand::thread_rng();
+            let token =
+                self.key_pair()
+                    .as_ref()
+                    .sign_with_rng(&mut rng, self.padding_scheme(), &digest)?;
+            Ok(token)
         })
     }
 }
 
 pub trait RSAPublicKeyLike {
+    type PaddingScheme: SignatureScheme;
+
     fn jwt_alg_name() -> &'static str;
     fn public_key(&self) -> &RSAPublicKey;
     fn key_id(&self) -> &Option<String>;
     fn set_key_id(&mut self, key_id: String);
-    fn hash() -> MessageDigest;
-    fn padding_scheme(&self) -> Padding;
+    fn hash(message: &[u8]) -> Vec<u8>;
+    fn padding_scheme(&self) -> Self::PaddingScheme;
+    fn padding_scheme_alt(&self) -> Option<Self::PaddingScheme>;
 
     fn verify_token<CustomClaims: Serialize + DeserializeOwned>(
         &self,
@@ -196,14 +205,22 @@ pub trait RSAPublicKeyLike {
             token,
             options,
             |authenticated, signature| {
-                let digest = Self::hash();
-                let pkey = PKey::from_rsa(self.public_key().as_ref().clone())?;
-                let mut verifier = Verifier::new(digest, &pkey)?;
-                verifier.update(authenticated.as_bytes())?;
-                if !(verifier
-                    .verify(signature)
-                    .map_err(|_| JWTError::InvalidSignature)?)
-                {
+                let digest = Self::hash(authenticated.as_bytes());
+                let mut verification_failed = self
+                    .public_key()
+                    .as_ref()
+                    .verify(self.padding_scheme(), &digest, signature)
+                    .is_err();
+                if verification_failed {
+                    if let Some(padding_scheme_alt) = self.padding_scheme_alt() {
+                        verification_failed = self
+                            .public_key()
+                            .as_ref()
+                            .verify(padding_scheme_alt, &digest, signature)
+                            .is_err();
+                    }
+                }
+                if verification_failed {
                     bail!(JWTError::InvalidSignature);
                 }
                 Ok(())
@@ -222,15 +239,22 @@ pub trait RSAPublicKeyLike {
             token,
             options,
             |authenticated, signature| {
-                let digest = Self::hash();
-                let pkey = PKey::from_rsa(self.public_key().as_ref().clone())?;
-                let mut verifier = Verifier::new(digest, &pkey)?;
-                verifier.update(authenticated.as_bytes())?;
-                if verifier
-                    .verify(&signature)
-                    .map_err(|_| JWTError::InvalidSignature)?
-                    == false
-                {
+                let digest = Self::hash(authenticated.as_bytes());
+                let mut verification_failed = self
+                    .public_key()
+                    .as_ref()
+                    .verify(self.padding_scheme(), &digest, signature)
+                    .is_err();
+                if verification_failed {
+                    if let Some(padding_scheme_alt) = self.padding_scheme_alt() {
+                        verification_failed = self
+                            .public_key()
+                            .as_ref()
+                            .verify(padding_scheme_alt, &digest, signature)
+                            .is_err();
+                    }
+                }
+                if verification_failed {
                     bail!(JWTError::InvalidSignature);
                 }
                 Ok(())
@@ -252,6 +276,8 @@ pub struct RS256PublicKey {
 }
 
 impl RSAKeyPairLike for RS256KeyPair {
+    type PaddingScheme = rsa::Pkcs1v15Sign;
+
     fn jwt_alg_name() -> &'static str {
         "RS256"
     }
@@ -273,12 +299,12 @@ impl RSAKeyPairLike for RS256KeyPair {
         Ok(())
     }
 
-    fn hash() -> MessageDigest {
-        MessageDigest::sha256()
+    fn hash(message: &[u8]) -> Vec<u8> {
+        sha2::Sha256::digest(message).to_vec()
     }
 
-    fn padding_scheme(&self) -> Padding {
-        Padding::PKCS1
+    fn padding_scheme(&self) -> rsa::Pkcs1v15Sign {
+        rsa::Pkcs1v15Sign::new::<sha2::Sha256>()
     }
 }
 
@@ -326,16 +352,22 @@ impl RS256KeyPair {
 }
 
 impl RSAPublicKeyLike for RS256PublicKey {
+    type PaddingScheme = rsa::Pkcs1v15Sign;
+
     fn jwt_alg_name() -> &'static str {
         "RS256"
     }
 
-    fn hash() -> MessageDigest {
-        MessageDigest::sha256()
+    fn hash(message: &[u8]) -> Vec<u8> {
+        sha2::Sha256::digest(message).to_vec()
     }
 
-    fn padding_scheme(&self) -> Padding {
-        Padding::PKCS1
+    fn padding_scheme(&self) -> rsa::Pkcs1v15Sign {
+        rsa::Pkcs1v15Sign::new::<sha2::Sha256>()
+    }
+
+    fn padding_scheme_alt(&self) -> Option<Self::PaddingScheme> {
+        None
     }
 
     fn public_key(&self) -> &RSAPublicKey {
@@ -391,11 +423,13 @@ impl RS256PublicKey {
     }
 
     pub fn sha1_thumbprint(&self) -> String {
-        Base64UrlSafeNoPadding::encode_to_string(SHA1::hash(&self.pk.to_der().unwrap())).unwrap()
+        Base64UrlSafeNoPadding::encode_to_string(sha1::Sha1::digest(&self.pk.to_der().unwrap()))
+            .unwrap()
     }
 
     pub fn sha256_thumbprint(&self) -> String {
-        Base64UrlSafeNoPadding::encode_to_string(SHA256::hash(&self.pk.to_der().unwrap())).unwrap()
+        Base64UrlSafeNoPadding::encode_to_string(sha2::Sha256::digest(&self.pk.to_der().unwrap()))
+            .unwrap()
     }
 }
 
@@ -414,6 +448,8 @@ pub struct RS512PublicKey {
 }
 
 impl RSAKeyPairLike for RS512KeyPair {
+    type PaddingScheme = rsa::Pkcs1v15Sign;
+
     fn jwt_alg_name() -> &'static str {
         "RS512"
     }
@@ -435,12 +471,12 @@ impl RSAKeyPairLike for RS512KeyPair {
         Ok(())
     }
 
-    fn hash() -> MessageDigest {
-        MessageDigest::sha512()
+    fn hash(message: &[u8]) -> Vec<u8> {
+        sha2::Sha512::digest(message).to_vec()
     }
 
-    fn padding_scheme(&self) -> Padding {
-        Padding::PKCS1
+    fn padding_scheme(&self) -> Self::PaddingScheme {
+        rsa::Pkcs1v15Sign::new::<sha2::Sha512>()
     }
 }
 
@@ -488,16 +524,22 @@ impl RS512KeyPair {
 }
 
 impl RSAPublicKeyLike for RS512PublicKey {
+    type PaddingScheme = rsa::Pkcs1v15Sign;
+
     fn jwt_alg_name() -> &'static str {
         "RS512"
     }
 
-    fn hash() -> MessageDigest {
-        MessageDigest::sha512()
+    fn hash(message: &[u8]) -> Vec<u8> {
+        sha2::Sha512::digest(message).to_vec()
     }
 
-    fn padding_scheme(&self) -> Padding {
-        Padding::PKCS1
+    fn padding_scheme(&self) -> Self::PaddingScheme {
+        rsa::Pkcs1v15Sign::new::<sha2::Sha512>()
+    }
+
+    fn padding_scheme_alt(&self) -> Option<Self::PaddingScheme> {
+        None
     }
 
     fn public_key(&self) -> &RSAPublicKey {
@@ -553,11 +595,13 @@ impl RS512PublicKey {
     }
 
     pub fn sha1_thumbprint(&self) -> String {
-        Base64UrlSafeNoPadding::encode_to_string(SHA1::hash(&self.pk.to_der().unwrap())).unwrap()
+        Base64UrlSafeNoPadding::encode_to_string(sha1::Sha1::digest(&self.pk.to_der().unwrap()))
+            .unwrap()
     }
 
     pub fn sha256_thumbprint(&self) -> String {
-        Base64UrlSafeNoPadding::encode_to_string(SHA256::hash(&self.pk.to_der().unwrap())).unwrap()
+        Base64UrlSafeNoPadding::encode_to_string(sha2::Sha256::digest(&self.pk.to_der().unwrap()))
+            .unwrap()
     }
 }
 
@@ -576,6 +620,8 @@ pub struct RS384PublicKey {
 }
 
 impl RSAKeyPairLike for RS384KeyPair {
+    type PaddingScheme = rsa::Pkcs1v15Sign;
+
     fn jwt_alg_name() -> &'static str {
         "RS384"
     }
@@ -597,12 +643,12 @@ impl RSAKeyPairLike for RS384KeyPair {
         Ok(())
     }
 
-    fn hash() -> MessageDigest {
-        MessageDigest::sha384()
+    fn hash(message: &[u8]) -> Vec<u8> {
+        sha2::Sha384::digest(message).to_vec()
     }
 
-    fn padding_scheme(&self) -> Padding {
-        Padding::PKCS1
+    fn padding_scheme(&self) -> Self::PaddingScheme {
+        rsa::Pkcs1v15Sign::new::<sha2::Sha384>()
     }
 }
 
@@ -650,16 +696,22 @@ impl RS384KeyPair {
 }
 
 impl RSAPublicKeyLike for RS384PublicKey {
+    type PaddingScheme = rsa::Pkcs1v15Sign;
+
     fn jwt_alg_name() -> &'static str {
         "RS384"
     }
 
-    fn hash() -> MessageDigest {
-        MessageDigest::sha384()
+    fn hash(message: &[u8]) -> Vec<u8> {
+        sha2::Sha384::digest(message).to_vec()
     }
 
-    fn padding_scheme(&self) -> Padding {
-        Padding::PKCS1
+    fn padding_scheme(&self) -> Self::PaddingScheme {
+        rsa::Pkcs1v15Sign::new::<sha2::Sha384>()
+    }
+
+    fn padding_scheme_alt(&self) -> Option<Self::PaddingScheme> {
+        None
     }
 
     fn public_key(&self) -> &RSAPublicKey {
@@ -715,11 +767,13 @@ impl RS384PublicKey {
     }
 
     pub fn sha1_thumbprint(&self) -> String {
-        Base64UrlSafeNoPadding::encode_to_string(SHA1::hash(&self.pk.to_der().unwrap())).unwrap()
+        Base64UrlSafeNoPadding::encode_to_string(sha1::Sha1::digest(&self.pk.to_der().unwrap()))
+            .unwrap()
     }
 
     pub fn sha256_thumbprint(&self) -> String {
-        Base64UrlSafeNoPadding::encode_to_string(SHA256::hash(&self.pk.to_der().unwrap())).unwrap()
+        Base64UrlSafeNoPadding::encode_to_string(sha2::Sha256::digest(&self.pk.to_der().unwrap()))
+            .unwrap()
     }
 }
 
@@ -738,6 +792,8 @@ pub struct PS256PublicKey {
 }
 
 impl RSAKeyPairLike for PS256KeyPair {
+    type PaddingScheme = rsa::Pss;
+
     fn jwt_alg_name() -> &'static str {
         "PS256"
     }
@@ -759,12 +815,12 @@ impl RSAKeyPairLike for PS256KeyPair {
         Ok(())
     }
 
-    fn hash() -> MessageDigest {
-        MessageDigest::sha256()
+    fn hash(message: &[u8]) -> Vec<u8> {
+        sha2::Sha256::digest(message).to_vec()
     }
 
-    fn padding_scheme(&self) -> Padding {
-        Padding::PKCS1_PSS
+    fn padding_scheme(&self) -> Self::PaddingScheme {
+        Self::PaddingScheme::new_with_salt::<sha2::Sha256>(256 / 8)
     }
 }
 
@@ -812,16 +868,22 @@ impl PS256KeyPair {
 }
 
 impl RSAPublicKeyLike for PS256PublicKey {
+    type PaddingScheme = rsa::Pss;
+
     fn jwt_alg_name() -> &'static str {
         "PS256"
     }
 
-    fn hash() -> MessageDigest {
-        MessageDigest::sha256()
+    fn hash(message: &[u8]) -> Vec<u8> {
+        sha2::Sha256::digest(message).to_vec()
     }
 
-    fn padding_scheme(&self) -> Padding {
-        Padding::PKCS1_PSS
+    fn padding_scheme(&self) -> Self::PaddingScheme {
+        Self::PaddingScheme::new_with_salt::<sha2::Sha256>(256 / 8)
+    }
+
+    fn padding_scheme_alt(&self) -> Option<Self::PaddingScheme> {
+        Some(Self::PaddingScheme::new::<sha2::Sha256>())
     }
 
     fn public_key(&self) -> &RSAPublicKey {
@@ -892,6 +954,8 @@ pub struct PS512PublicKey {
 }
 
 impl RSAKeyPairLike for PS512KeyPair {
+    type PaddingScheme = rsa::Pss;
+
     fn jwt_alg_name() -> &'static str {
         "PS512"
     }
@@ -913,12 +977,12 @@ impl RSAKeyPairLike for PS512KeyPair {
         Ok(())
     }
 
-    fn hash() -> MessageDigest {
-        MessageDigest::sha512()
+    fn hash(message: &[u8]) -> Vec<u8> {
+        sha2::Sha512::digest(message).to_vec()
     }
 
-    fn padding_scheme(&self) -> Padding {
-        Padding::PKCS1_PSS
+    fn padding_scheme(&self) -> Self::PaddingScheme {
+        Self::PaddingScheme::new_with_salt::<sha2::Sha512>(512 / 8)
     }
 }
 
@@ -966,16 +1030,22 @@ impl PS512KeyPair {
 }
 
 impl RSAPublicKeyLike for PS512PublicKey {
+    type PaddingScheme = rsa::Pss;
+
     fn jwt_alg_name() -> &'static str {
         "PS512"
     }
 
-    fn hash() -> MessageDigest {
-        MessageDigest::sha512()
+    fn hash(message: &[u8]) -> Vec<u8> {
+        sha2::Sha512::digest(message).to_vec()
     }
 
-    fn padding_scheme(&self) -> Padding {
-        Padding::PKCS1_PSS
+    fn padding_scheme(&self) -> Self::PaddingScheme {
+        Self::PaddingScheme::new_with_salt::<sha2::Sha512>(512 / 8)
+    }
+
+    fn padding_scheme_alt(&self) -> Option<Self::PaddingScheme> {
+        Some(Self::PaddingScheme::new::<sha2::Sha512>())
     }
 
     fn public_key(&self) -> &RSAPublicKey {
@@ -1031,11 +1101,13 @@ impl PS512PublicKey {
     }
 
     pub fn sha1_thumbprint(&self) -> String {
-        Base64UrlSafeNoPadding::encode_to_string(SHA1::hash(&self.pk.to_der().unwrap())).unwrap()
+        Base64UrlSafeNoPadding::encode_to_string(sha1::Sha1::digest(&self.pk.to_der().unwrap()))
+            .unwrap()
     }
 
     pub fn sha256_thumbprint(&self) -> String {
-        Base64UrlSafeNoPadding::encode_to_string(SHA256::hash(&self.pk.to_der().unwrap())).unwrap()
+        Base64UrlSafeNoPadding::encode_to_string(sha2::Sha256::digest(&self.pk.to_der().unwrap()))
+            .unwrap()
     }
 }
 
@@ -1054,6 +1126,8 @@ pub struct PS384PublicKey {
 }
 
 impl RSAKeyPairLike for PS384KeyPair {
+    type PaddingScheme = rsa::Pss;
+
     fn jwt_alg_name() -> &'static str {
         "PS384"
     }
@@ -1075,12 +1149,12 @@ impl RSAKeyPairLike for PS384KeyPair {
         Ok(())
     }
 
-    fn hash() -> MessageDigest {
-        MessageDigest::sha384()
+    fn hash(message: &[u8]) -> Vec<u8> {
+        sha2::Sha384::digest(message).to_vec()
     }
 
-    fn padding_scheme(&self) -> Padding {
-        Padding::PKCS1_PSS
+    fn padding_scheme(&self) -> Self::PaddingScheme {
+        Self::PaddingScheme::new_with_salt::<sha2::Sha384>(384 / 8)
     }
 }
 
@@ -1128,16 +1202,22 @@ impl PS384KeyPair {
 }
 
 impl RSAPublicKeyLike for PS384PublicKey {
+    type PaddingScheme = rsa::Pss;
+
     fn jwt_alg_name() -> &'static str {
         "PS384"
     }
 
-    fn hash() -> MessageDigest {
-        MessageDigest::sha384()
+    fn hash(message: &[u8]) -> Vec<u8> {
+        sha2::Sha384::digest(message).to_vec()
     }
 
-    fn padding_scheme(&self) -> Padding {
-        Padding::PKCS1_PSS
+    fn padding_scheme(&self) -> Self::PaddingScheme {
+        Self::PaddingScheme::new_with_salt::<sha2::Sha384>(384 / 8)
+    }
+
+    fn padding_scheme_alt(&self) -> Option<Self::PaddingScheme> {
+        Some(Self::PaddingScheme::new::<sha2::Sha384>())
     }
 
     fn public_key(&self) -> &RSAPublicKey {
@@ -1193,10 +1273,12 @@ impl PS384PublicKey {
     }
 
     pub fn sha1_thumbprint(&self) -> String {
-        Base64UrlSafeNoPadding::encode_to_string(SHA1::hash(&self.pk.to_der().unwrap())).unwrap()
+        Base64UrlSafeNoPadding::encode_to_string(sha1::Sha1::digest(&self.pk.to_der().unwrap()))
+            .unwrap()
     }
 
     pub fn sha256_thumbprint(&self) -> String {
-        Base64UrlSafeNoPadding::encode_to_string(SHA256::hash(&self.pk.to_der().unwrap())).unwrap()
+        Base64UrlSafeNoPadding::encode_to_string(sha2::Sha256::digest(&self.pk.to_der().unwrap()))
+            .unwrap()
     }
 }
